All -- Refactor -- Better ui

+++ back/Core/Interfaces/Hubs/IRoomHub.cs
@@ -5,5 +5,5 @@ namespace Core.Interfaces.Hubs;
 public interface IRoomHub
 {
     public Task UpdateVideoState(string idRoom, RoomState state);
-    public Task SeekTime(string idRoom, long time);
+    public Task SeekTime(string idRoom, double time);
 }
